Add listener for the queues
